Rename shared to _shared
