+++ src/azure-cli-core/azure/cli/core/commands/client_factory.py
@@ -10,7 +10,7 @@ import azure.cli.core._debug as _debug
 import azure.cli.core.azlogging as azlogging
 from azure.cli.core.util import CLIError
 from azure.cli.core.application import APPLICATION
-from azure.cli.core.profiles.shared import get_client_class
+from azure.cli.core.profiles._shared import get_client_class
 from azure.cli.core.profiles import get_api_version, get_sdk, ResourceType
 
 logger = azlogging.get_az_logger(__name__)
