+++ src/azure-cli-core/azure/cli/core/profiles/_shared.py
